+++ game_patch/hud/multi_hud_chat.cpp
@@ -56,6 +56,10 @@ void multi_hud_render_chat()
         fade_out = rf::chat_fade_out_timer.time_until() / 750.0f;
     }
 
+    if (g_remote_server_cfg_popup.is_active()) {
+        return;
+    }
+
     int chatbox_font = hud_get_default_font();
     int clip_w = rf::gr::clip_width();
     int font_h = rf::gr::get_font_height(chatbox_font);
@@ -82,8 +86,7 @@ void multi_hud_render_chat()
         }
         int x = box_x + border + 6;
 
-        int name_w, name_h;
-        rf::gr::get_string_size(&name_w, &name_h, msg.name.c_str(), -1, chatbox_font);
+        const auto [name_w, name_h] = rf::gr::get_string_size(msg.name, chatbox_font);
         if (msg.color_id == 0 || msg.color_id == 1) {
             if (msg.color_id == 0) {
                 rf::gr::set_color(227, 48, 47, text_alpha);
@@ -123,6 +126,10 @@ CodeInjection multi_hud_add_chat_line_max_width_injection{
 
 void multi_hud_render_chat_inputbox(rf::String::Pod label_pod, rf::String::Pod msg_pod)
 {
+    if (g_remote_server_cfg_popup.is_active()) {
+        return;
+    }
+
     // Note: POD has to be used here because of differences between compilers ABI
     rf::String label{label_pod};
     rf::String msg{msg_pod};
@@ -144,10 +151,8 @@ void multi_hud_render_chat_inputbox(rf::String::Pod label_pod, rf::String::Pod m
     int input_box_x = (clip_w - box_w) / 2; // 157
 
     rf::String msg_shortened{msg};
-    int msg_w, msg_h;
-    rf::gr::get_string_size(&msg_w, &msg_h, msg_shortened.c_str(), -1, chatbox_font);
-    int label_w, label_h;
-    rf::gr::get_string_size(&label_w, &label_h, label.c_str(), -1, chatbox_font);
+    auto [msg_w, msg_h] = rf::gr::get_string_size(msg_shortened, chatbox_font);
+    const auto [label_w, label_h] = rf::gr::get_string_size(label, chatbox_font);
     int cursor_w = g_big_chatbox ? 10 : 5;
     int cursor_h = g_big_chatbox ? 2 : 1;
     int max_msg_w = content_w - cursor_w - 7 - label_w;
@@ -157,7 +162,7 @@ void multi_hud_render_chat_inputbox(rf::String::Pod label_pod, rf::String::Pod m
     }
     while (msg_w > max_msg_w) {
         msg_shortened = msg_shortened.substr(1, -1);
-        rf::gr::get_string_size(&msg_w, &msg_h, msg_shortened.c_str(), -1, chatbox_font);
+        std::tie(msg_w, msg_h) = rf::gr::get_string_size(msg_shortened, chatbox_font);
     }
 
     rf::gr::set_color(255, 255, 255, rf::scoreboard_visible ? 255 : chatbox_border_alpha);
@@ -235,6 +240,15 @@ ConsoleCommand2 mute_player_cmd{
     "Mutes a single player in multiplayer chat",
 };
 
+FunHook<void(int)> multi_chat_say_show_hook{
+    0x00444A80,
+    [] (const int is_team_chat) {
+        if (!g_remote_server_cfg_popup.is_active()) {
+            multi_chat_say_show_hook.call_target(is_team_chat);
+        }
+    },
+};
+
 void multi_hud_chat_apply_patches()
 {
     // Fix game beeping every frame if chat input buffer is full
@@ -258,6 +272,8 @@ void multi_hud_chat_apply_patches()
 
     // Do not strip '%' characters from chat messages
     write_mem<u8>(0x004785FD, asm_opcodes::jmp_rel_short);
+
+    multi_chat_say_show_hook.install();
 }
 
 void multi_hud_chat_set_big(bool is_big)